% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bcSubSampleSerial}
\alias{bcSubSampleSerial}
\title{Serial biclustering}
\usage{
bcSubSampleSerial(geneDf, nSamples = 100, lam, propSample = 0.6,
  lam.lwr = 3.5, clustOptions = list())
}
\arguments{
\item{nSamples}{integer denoting the number of permutations to perform}

\item{lam}{regularization parameter for conditions (maximum number of conditions allows in a bicluster)}

\item{expression}{matrix with features on rows and conditions on columns}
}
\description{
Uses one core to perform SCCAB
}

