% Generated by roxygen2 (4.0.2): do not edit by hand
\name{biClustMax.optim}
\alias{biClustMax.optim}
\title{Optimizing the biclustering function}
\usage{
biClustMax.optim(X, Y, d.start, lam, verbose = TRUE, lam.lwr = 3.5,
  clustOptions = list())
}
\arguments{
\item{X}{matrix with rows representing genes and columns representing conditions (n x k)}

\item{Y}{matrix with rows representing genes and columns representing conditions (n x k)}

\item{d.start}{vector of dimension k}

\item{lam}{maximum cluster size}
}
\value{
A list containing maximum values of a, b, d
}
\description{
One iteration of biclustering maximization. Takes the current values of
d, and if successful returns one iteration.
}

