% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/post_process.R
\name{count_index}
\alias{count_index}
\title{count indeces}
\usage{
count_index(idx)
}
\arguments{
\item{idx}{a vector of indices (usually integer) with repeating values}
}
\value{
a data.frame with columns "index" referring to the index and
"count", the count of the index
}
\description{
count indeces
}

