% Generated by roxygen2 (4.0.2): do not edit by hand
\name{detect_d_maximizer}
\alias{detect_d_maximizer}
\title{Detect the conditions maximizer to use}
\usage{
detect_d_maximizer(opt)
}
\arguments{
\item{opt}{a \code{string} either "lasso" or "timeseries". If
a \code{function}, simply use the \code{function}. Does not check the return
type if \code{opt} is a \code{function}.}
}
\value{
a function to maximize the condition coefficients
}
\description{
Given a string or a function, return a function to maximize the condition
coefficients.
}

