% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/max_features.R
\name{features_max_fscca}
\alias{features_max_fscca}
\title{Maximize the features using fscca}
\usage{
features_max_fscca(X, Y, d, lam_a, lam_b)
}
\arguments{
\item{X}{matrix of size n x p}

\item{Y}{matrix of size n x p}

\item{d}{vector of length p denoting inclusion}

\item{lam_a}{a vector of parameters to regularize a}

\item{lam_b}{a vector of parameters to regularize b}
}
\value{
a list with 'a' and 'b' solutions from SCCA
}
\description{
Wrapper to maximize features using fscca
}

