% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_low_expression}
\alias{filter_low_expression}
\title{Filter out rows and columns with low expression}
\usage{
filter_low_expression(data, min_exp, max_cols)
}
\arguments{
\item{data}{the matrix to filter}

\item{min_exp}{the minimum expression in each row}

\item{max_cols}{the maximum number of cols that can have \code{min_exp}}
}
\description{
Filter out rows and columns with low expression
}

