% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/benchmark.R
\name{iid_gaussian_block}
\alias{iid_gaussian_block}
\title{Generate a matrix that is iid normal}
\usage{
iid_gaussian_block(nrows, ncols, mean, sd)
}
\arguments{
\item{nrows}{the number of rows}

\item{ncols}{the number of cols}

\item{mean}{the mean of each sample}

\item{sd}{the standard devian of each sample}
}
\value{
a \code{matrix} of size nrows * ncols with iid samples from N(mean,
sd)
}
\description{
Generate a matrix that is iid normal
}

