% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/benchmark.R
\name{insert_matrix}
\alias{insert_matrix}
\title{Insert a submatrix into a matrix}
\usage{
insert_matrix(sub_matrix, row_start, col_start, bg_matrix)
}
\arguments{
\item{sub_matrix}{the matrix to insert}

\item{row_start}{the row to insert the submatrix into the background matrix}

\item{col_start}{the column to insert the submatrix into the background matrix}

\item{bg_matrix}{the background matrix}
}
\value{
the matrix \code{bg_matrix} with \code{sub_matrix} inserted at
position \code{row_start} and \code{col_start}.
}
\description{
Take a matrix that is smaller and insert it into a particular position in
a larger matrix
}

