% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lasso_max_d}
\alias{lasso_max_d}
\title{Maximize the d vector in biclustering}
\usage{
lasso_max_d(X, Y, a, b, s)
}
\arguments{
\item{X}{a matrix with features on rows}

\item{Y}{a matrix with features on rows}

\item{a}{fixed estimates of feature coefficients on X}

\item{b}{fixed estimates of feature coefficients on Y}

\item{s}{a regularization parameter such that \eqn{s > 0} and \eqn{d^T q d  <= s}}
}
\value{
a valid maximum
}
\description{
Find the solution to the quadratic program:
\eqn{q = a^T X b^T Y}
\eqn{f(d) = d^2 * q}
\eqn{s.t. ||d|| <= s, 0 <= d_i <= 1}
}

