% Generated by roxygen2 (4.0.2): do not edit by hand
\name{maximizeOneSplit}
\alias{maximizeOneSplit}
\title{Driver function to find solution for}
\usage{
maximizeOneSplit(geneDf, lam, epsA = 0.001, epsB = 0.001, epsD = 0.01,
  maxIt = 100, lam.lwr, clustOptions = list())
}
\description{
Driver function to find solution for
}

