% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/benchmark.R
\name{mvn_block}
\alias{mvn_block}
\title{Generate a multivariate normal block}
\usage{
mvn_block(nrows = 30, ncols = 20, min = 0.5, max = 0.8)
}
\arguments{
\item{nrows}{the number of rows}

\item{ncols}{the number of cols}

\item{min}{the min number to pass to \link{\code{random_pos_def_mat}}}

\item{max}{the max number to pass to \link{\code{random_pos_def_mat}}}
}
\value{
a multivariate normal with dimensions nrows * ncols
}
\description{
Generate a multivariate normal block
}

