% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/benchmark.R
\name{permute_mat}
\alias{permute_mat}
\title{Permute a matrix}
\usage{
permute_mat(mat)
}
\arguments{
\item{mat}{any object (usually a \code{matrix}) that accepts functions \code{nrow} and \code{ncol}}
}
\value{
a list with named elements:
\itemize{
 \item{"row_order"}{the order of the rows}
 \item{col_order}{the order of the columns}
 \item{permuted}{the permuted matrix}
 \item{original}{the original matrix}
}
}
\description{
Randomly permute the rows and columns of a matrix.
}
\seealso{
\code{\link{translate_idx}}
}

