% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bcPlots.R
\name{plotClusterExpression}
\alias{plotClusterExpression}
\title{Plot the expression only in the bicluster}
\usage{
plotClusterExpression(exMat, clust)
}
\arguments{
\item{exMat}{the expression matrix}

\item{clust}{the post-process clustering output}
}
\value{
a gpplot object
}
\description{
Given some solution from a post-process method, plot only those rows and
columns
}

