% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/post_process.R
\name{pps_sol}
\alias{pps_sol}
\title{Constructor for biclustering result}
\usage{
pps_sol(rows, cols)
}
\arguments{
\item{rows}{an \code{integer} vector of rows in the bicluster}

\item{cols}{an \code{integer} vector of columns in the bicluster}
}
\value{
A \code{pps_sol} object, which is essentially a named list with
\describe{
 \item{row_idx}{The indices of the rows}
 \item{col_idx}{The indices of the columns}
}
}
\description{
Given a vector of rows and a vector of columns, turn it into a formal SCCAB
\code{pps_sol} result.
}
\seealso{
\code{\link{pps}}, \code{\link{pps_sub}}, \code{\link{get_pps}}
}

