% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/params.R
\name{print.sccab_params}
\alias{print.sccab_params}
\title{Print sccab_params object}
\usage{
\method{print}{sccab_params}(obj)
}
\arguments{
\item{\code{obj}}{an object of type \code{sccab_params}}
}
\value{
the original \code{obj}
}
\description{
Print \code{sccab_params} object and returns it \emph{invisibly} (via
\code{\link{invisible}(x)}).
}

