% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biclustering.R
\name{print.sccab_result}
\alias{print.sccab_result}
\title{Printing a sccab_result}
\usage{
\method{print}{sccab_result}(obj)
}
\arguments{
\item{obj}{the \code{sccab_result} object to be printed}
}
\value{
returns \code{obj} using \code{\link{invisible}(obj)}
}
\description{
If \code{sccab} is called (the non-subsampled version) and
\code{!is.na(prop)} in \code{sccab_params}, it will still print
a subsampling proportion, even though no subsampling occured.
}

