% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/benchmark.R
\name{random_pos_def_mat}
\alias{random_pos_def_mat}
\title{Generate a random posisitive definite matrix}
\usage{
random_pos_def_mat(nvars = 10, min = 0.5, max = 0.8)
}
\arguments{
\item{nvars}{the number of variables}

\item{min}{the minimum number to generate the covariance from}

\item{max}{the maximum number to generate the covariance from}
}
\value{
a random positive definite matrix with diagonal 1
}
\description{
Generate a random posisitive definite matrix
}

