% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biclustering.R
\name{sccab}
\alias{sccab}
\title{SCCAB(iclustering) (with all data)}
\usage{
sccab(exp_mat, params)
}
\arguments{
\item{exp_mat}{matrix with features on rows and conditions on columns}

\item{params}{a \code{sccab_params} object from \code{sccab_params()}}
}
\value{
a \code{sccab_res} object
}
\description{
Given a set of features, will find the most "dominant" bicluster. Works in
parallel using library \code{\link[multicore]{multicore}}. To set the number of cores
used, set \code{options(cores = N)}.
}

