% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sccab_params}
\alias{sccab_params}
\title{Set SCCAB(biclustering) parameters}
\usage{
sccab_params(d_upr, n_samp = 100, d_lwr = 3, ab_lam = c(0.5, 3, 10),
  optim_fun = "lasso", prop = NA, parallel = FALSE, verbose = TRUE)
}
\arguments{
\item{d_upr}{regularization parameter for conditions (maximum number of conditions allows in a bicluster)}

\item{n_samp}{integer denoting the number of permutations to perform}

\item{d_lwr}{regularization parameter for conditions (minimum number of conditions allows in a bicluster)}

\item{ab_lam}{vector of regularization parameters for expression features.}

\item{optim_fun}{"lasso" or "timeseries" or a user-supplied function. NOTE: there are no checks for validity of the user-supplied function.}

\item{prop}{proportion to subsample}

\item{parallel}{if TRUE use parallel::mclapply, else use lapply}

\item{verbose}{if TRUE print messages}
}
\value{
A sccab_params object to be used in sccab or sccab_subsample
}
\description{
Call this function before calling any sccab function. The result is used in
the sccab functions to define behavior.
}

