% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biclustering.R
\name{sccab_subsample}
\alias{sccab_subsample}
\title{SCCAB(iclustering) with subsampling}
\usage{
sccab_subsample(exp_mat, params)
}
\arguments{
\item{exp_mat}{matrix with features on rows and conditions on columns}

\item{params}{a sccab_params object from sccab_params()}
}
\value{
a sccab_res object
}
\description{
Given a set of features, will find the most "dominant" bicluster. Works in
parallel using library "multicore." To set the number of cores used, set
options(mc.cores = N).
}

