% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{split_evenly}
\alias{split_evenly}
\title{Randomly split a set of indices}
\usage{
split_evenly(n)
}
\arguments{
\item{n}{the number of indices}
}
\value{
a list with two disjoint sets of indices from 1 to n.
}
\description{
Randomly splits indices if there are an even number. If there is an odd
number, it will stop().
}

