% Generated by roxygen2 (4.0.2): do not edit by hand
\name{splitEvenly}
\alias{splitEvenly}
\title{Randomly splits indices if there are an even number}
\usage{
splitEvenly(n)
}
\arguments{
\item{n}{the number of indices}
}
\value{
a list with two disjoint sets of indices from 1 to n.
}
\description{
Randomly splits indices if there are an even number
}

