% Generated by roxygen2 (4.0.2): do not edit by hand
\name{biClustMax.optim}
\alias{biClustMax.optim}
\title{One iteration of biclustering maximization. Takes the current values of
d, a, b and if successful returns one iteration}
\usage{
biClustMax.optim(X, Y, d.start, a, b, lam, verbose = TRUE, optim.max = 4,
  lam.lwr = 3.5, clustOptions = list())
}
\arguments{
\item{a}{vector of dimension n}

\item{b}{vector of dimension n}

\item{lam}{maximum cluster size}

\item{xDf}{matrix with rows representing genes and columns representing conditions (n x k)}

\item{yDf}{matrix with rows representing genes and columns representing conditions (n x k)}

\item{d}{vector of dimension k}
}
\value{
A list containing maximum values of a, b, d
}
\description{
One iteration of biclustering maximization. Takes the current values of
d, a, b and if successful returns one iteration
}

