% Generated by roxygen2 (4.0.2): do not edit by hand
\name{biClustMax.optim}
\alias{biClustMax.optim}
\title{Optimizing the biclustering function}
\usage{
biClustMax.optim(X, Y, d_init, ab_lam, d_lwr, d_upr, d_max_fun)
}
\arguments{
\item{X}{matrix with rows representing genes and columns representing conditions (n x k)}

\item{Y}{matrix with rows representing genes and columns representing conditions (n x k)}

\item{d_init}{vector of dimension k}

\item{ab_lam}{regularization on expression features}

\item{ab_lam}{regularization on expression features}
}
\value{
A list containing maximum values of a, b, d
}
\description{
One iteration of biclustering maximization. Takes the current values of
d, and if successful returns one iteration.
}

