% Generated by roxygen2 (4.0.2): do not edit by hand
\name{count_index}
\alias{count_index}
\title{count indeces}
\usage{
count_index(idx)
}
\arguments{
\item{idx}{a vector of indices (usually integer) with repeating values}
}
\value{
a data.frame with columns "index" referring to the index and
"count", the count of the index
}
\description{
count indeces
}

