% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_low_expression}
\alias{filter_low_expression}
\title{Filter out rows and columns with low expression}
\usage{
filter_low_expression(data, min_exp, min_percent)
}
\arguments{
\item{data}{the matrix to filter}

\item{min_exp}{the minimum expression in each row}

\item{min_percent}{the minimum percentage of the rows}
}
\description{
Filter out rows and columns with low expression
}

