% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggPlotExpression}
\alias{ggPlotExpression}
\title{Heatmap of expression}
\usage{
ggPlotExpression(exMat, cluster = NULL, clustRows = T, clustCols = T,
  rowNames = F, colNames = T)
}
\arguments{
\item{exMat}{the expression matrix}

\item{cluster}{either the cluster results to group the rows/columns in the
biclustering, or NULL if there are no results}

\item{clustRows}{if TRUE, cluster the rows by hierarchical clustering.}

\item{clustCols}{if TRUE, cluster the columns by hierarchical clustering.}

\item{rowNames}{if TRUE, print the row names on the plot}

\item{colNames}{if TRUE, print the column names on the plot}
}
\value{
a ggplot object
}
\description{
Plot all of the points in an expression matrix
}

