% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggPlotSSSolution}
\alias{ggPlotSSSolution}
\title{Plot the parallel solution}
\usage{
ggPlotSSSolution(parSol, clust = NA, fBase = NA, rowNames = NA,
  colNames = NA)
}
\arguments{
\item{parSol}{the solution}

\item{clust}{if you've post-process clustered the clustered object denoting
which rows/columns exist in the bicluster}

\item{fBase}{a string for the file name}

\item{rowNames}{if not NA, the rownames to use for expression features}

\item{colNames}{if not NA, a character vector of column (condition) names}
}
\value{
A list with two ggplot objects
}
\description{
Plot a solution from either bcSubSampleSerial or bcSubSamplePar
}

