% Generated by roxygen2 (4.0.2): do not edit by hand
\name{max_one_split}
\alias{max_one_split}
\title{Maximize one split of the SCCAB algorithm}
\usage{
max_one_split(exp_mat, params, epsA = 0.001, epsB = 0.001, epsD = 0.01,
  maxIt = 100)
}
\description{
Maximize one split of the SCCAB algorithm
}

