% Generated by roxygen2 (4.0.2): do not edit by hand
\name{maximizeOneSplit}
\alias{maximizeOneSplit}
\title{Maximize one split of the SCCAB algorithm}
\usage{
maximizeOneSplit(exp_mat, lam, epsA = 0.001, epsB = 0.001, epsD = 0.01,
  maxIt = 100, lam.lwr, clustOptions = list())
}
\description{
Maximize one split of the SCCAB algorithm
}

