% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sccab}
\alias{sccab}
\title{SCCAB(iclustering) (with all data)}
\usage{
sccab(exp_mat, params)
}
\arguments{
\item{exp_mat}{matrix with features on rows and conditions on columns}

\item{params}{a sccab_params object from sccab_params()}
}
\value{
a sccab_res object
}
\description{
Given a set of features, will find the most "dominant" bicluster. Works in
parallel using library "multicore." To set the number of cores used, set
options(cores = N).
}

