% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sccab}
\alias{sccab}
\title{SCCAB(iclustering) (with all data)}
\usage{
sccab(exp_mat, lam_upr, n_samples = 100, lam_lwr = 3.5, parallel = FALSE,
  clust_opt = list())
}
\arguments{
\item{exp_mat}{matrix with features on rows and conditions on columns}

\item{lam_upr}{regularization parameter for conditions (maximum number of conditions allows in a bicluster)}

\item{n_samples}{integer denoting the number of permutations to perform}

\item{lam_lwr}{the lower boundary lambda (minimum number of conditions)}

\item{parallel}{if TRUE, use parallel::mclapply instead of lapply}

\item{clust_opt}{a list of additional cluster options}
}
\description{
Given a set of features, will find the most "dominant" bicluster. Works in
parallel using library "multicore." To set the number of cores used, set
options(cores = N).
}

