% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sccab}
\alias{sccab}
\title{SCCAB(iclustering) (with all data)}
\usage{
sccab(exp_mat, nSamples = 100, lam, lam.lwr = 3.5, parallel = FALSE,
  clustOptions = list())
}
\arguments{
\item{exp_mat}{matrix with features on rows and conditions on columns}

\item{nSamples}{integer denoting the number of permutations to perform}

\item{lam}{regularization parameter for conditions (maximum number of conditions allows in a bicluster)}

\item{parallel}{if TRUE, use parallel::mclapply instead of lapply}

\item{lam_lwr}{the lower boundary lambda (minimum number of conditions)}

\item{parallel}{if TRUE use parallel::mclapply, else use lapply}

\item{clust_opt}{a list of additional cluster options}
}
\description{
Given a set of features, will find the most "dominant" bicluster. Works in
parallel using library "multicore." To set the number of cores used, set
options(cores = N).
}

