% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/benchmark.R
\name{sccab_sims}
\alias{sccab_sims}
\title{Compute sccab over a bunch of simulations}
\usage{
sccab_sims(sccab_sim, params)
}
\arguments{
\item{params}{a \code{sccab_params} object}

\item{data}{A named list containing \code{sims}, a list of simulations}

\item{sccab_func}{currently not used, a function to call sccab_with}
}
\value{
a list of \code{sccab} results
}
\description{
Compute sccab over a bunch of simulations and get back their results.
}

