% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sccab_subsample}
\alias{sccab_subsample}
\title{SCCAB(iclustering) with subsampling}
\usage{
sccab_subsample(exp_mat, lam_upr, n_samp = 100, prop = 0.6, lam_lwr = 3.5,
  parallel = FALSE, clust_opt = list())
}
\arguments{
\item{exp_mat}{matrix with features on rows and conditions on columns}

\item{params}{a sccab_params object from the sccab_params function}
}
\description{
Given a set of features, will find the most "dominant" bicluster. Works in
parallel using library "multicore." To set the number of cores used, set
options(mc.cores = N).
}

